move handlers of primary to separate module
