+++ narwhal/primary/src/lib.rs
@@ -29,6 +29,7 @@ mod metrics;
 #[cfg(test)]
 #[path = "tests/certificate_tests.rs"]
 mod certificate_tests;
+mod handlers;
 
 pub use crate::{
     block_remover::BlockRemover,
