+++ narwhal/primary/src/tests/primary_tests.rs
@@ -1,6 +1,6 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-use super::{NetworkModel, Primary, PrimaryReceiverHandler, CHANNEL_CAPACITY};
+use super::{handlers::PrimaryReceiverHandler, NetworkModel, Primary, CHANNEL_CAPACITY};
 use crate::{
     common::create_db_stores,
     metrics::{PrimaryChannelMetrics, PrimaryMetrics},
